# Add cache for Basic Authentication