--- a/bundles/org.openhab.core.io.rest.auth/src/main/java/org/openhab/core/io/rest/auth/internal/AuthFilter.java
+++ b/bundles/org.openhab.core.io.rest.auth/src/main/java/org/openhab/core/io/rest/auth/internal/AuthFilter.java
@@ -147,8 +147,7 @@ public class AuthFilter implements ContainerRequestFilter {
         userRegistry.removeRegistryChangeListener(userRegistryListener);
     }
 
-    @Nullable
-    private String getCacheKey(String credentials) {
+    private @Nullable String getCacheKey(String credentials) {
         try {
             final MessageDigest md = MessageDigest.getInstance("SHA-256");
             md.update(RANDOM_BYTES);