--- a/bundles/org.openhab.core.io.rest.auth/src/main/java/org/openhab/core/io/rest/auth/internal/ExpiringUserSecurityContextCache.java
+++ b/bundles/org.openhab.core.io.rest.auth/src/main/java/org/openhab/core/io/rest/auth/internal/ExpiringUserSecurityContextCache.java
@@ -71,10 +71,13 @@ public class ExpiringUserSecurityContextCache {
     private @Nullable Entry getEntry(String key) {
         Entry entry = entryMap.get(key);
         if (entry != null) {
-            long entryAge = System.currentTimeMillis() - entry.timestamp;
+            final long curTimeMillis = System.currentTimeMillis();
+            long entryAge = curTimeMillis - entry.timestamp;
             if (entryAge < 0 || entryAge >= keepPeriod) {
                 entryMap.remove(key);
                 entry = null;
+            } else {
+                entry.timestamp = curTimeMillis;
             }
         }
         return entry;