--- a/bundles/org.openhab.core.io.rest.auth/src/main/java/org/openhab/core/io/rest/auth/internal/ExpiringUserSecurityContextCache.java
+++ b/bundles/org.openhab.core.io.rest.auth/src/main/java/org/openhab/core/io/rest/auth/internal/ExpiringUserSecurityContextCache.java
@@ -84,7 +84,7 @@ public class ExpiringUserSecurityContextCache {
     }
 
     static class Entry {
-        final public long timestamp;
+        public long timestamp;
         final public UserSecurityContext value;
 
         Entry(long timestamp, UserSecurityContext value) {