--- a/bundles/org.openhab.core.io.rest.auth/src/main/resources/OH-INF/config/config.xml
+++ b/bundles/org.openhab.core.io.rest.auth/src/main/resources/OH-INF/config/config.xml
@@ -11,7 +11,7 @@
 			<description>Allow the use of Basic authentication to access protected API resources, in addition to access tokens
 				and API tokens.</description>
 		</parameter>
-		<parameter name="cacheExpiration" type="integer" min="1">
+		<parameter name="cacheExpiration" type="integer" min="0">
 			<advanced>true</advanced>
 			<label>Cache Expiration Time</label>
 			<default>6</default>
@@ -20,6 +20,7 @@
 				authorization.
 				The entries in the cache expire after a while in order to not keep credentials in memory indefinitely.
 				This value defines the expiration time in hours.
+				Set it to 0 for disabling the cache.
 			</description>
 		</parameter>
 		<parameter name="implicitUserRole" type="boolean" required="false">