This PR adds a small cache for calculating the auth hashes which aims to fix the performance issues introduced with OH3.
The cache does **not** store plain text passwords and is limited to a certain size.

Fixes https://github.com/openhab/openhab-core/issues/1899