@spacemanspiff2007 2021-01-12
This brings OH3 close to OH2 levels (~10%)

```
Benchmark item creation
Updated 300 item definitions in 2.598s --> 115.467 updates per sec

Benchmark value update creation
Wait: 0.000
Benchmark duration: 2.009s --> 149.319 updates per sec
Pings (13): min: 9.0 max: 26.0 median: 24.0
```