@Pedals2Paddles 2021-01-12
So this will fix logging into OH3 main UI with a username/password after using `bundle:stop org.openhab.core.io.rest.auth`?