@spacemanspiff2007 2021-01-12
No, this will make that you don't need ``bundle:stop org.openhab.core.io.rest.auth`` any more.
Which then will make the main UI work again.