@Pedals2Paddles 2021-01-12
Oh that's even better!  Is this a workaround for the problem, or *the* cause and solution to the problem?  I'll take either as a victory at this point of course.