@spacemanspiff2007 2021-01-13
Imho the PR tests breaks because of something else:
```
[INFO] openHAB Core :: Integration Tests :: IO REST Core Tests SUCCESS [ 18.996 s]
[INFO] openHAB Core :: Integration Tests :: Model Core Tests FAILURE [  5.271 s]
```