@Pedals2Paddles 2021-01-17
Bumping for maintainer attention. I truly feel this should be considered a critical fix. Openhab is **unusable** through NGINX with this auth problem.