@splatch 2021-01-18
Isn't issue is caused by pbkf2 passwords which needs to be calculated for each and every request?

IMHO Place which gets fixed by this commit is where you see symptoms but not where root cause is.