@spacemanspiff2007 2021-01-19
This commit adds a cache so the password hashes don't have to be calculated for every request any more.
The root cause for the performance issue is the long and multiple hash times but I don't know how and if this should be tackled,

@splatch 
I am not sure I understand correctly what you are trying to say. Could you elaborate a little bit further?