@splatch 2021-01-19
> I am not sure I understand correctly what you are trying to say. Could you elaborate a little bit further?

My suggestion is to consider that functionality into ManagedUserProvider/Registry which does password checks. This is specific implementation which is affected by hashing algo.