@spacemanspiff2007 2021-01-19
Yes - that would be even better. 
However maybe let's get this merged first so OH3 is usable again and then we can tackle the proper implementation.
Since I am not a Java developer it took me quite some time to find this implementation and I'd like to migrate to OH3, too so I can provide better support for the HABApp users.