@Pedals2Paddles 2021-01-20
Well according to the maintainer for openhabion, there's already a solution for this. But he's currently refusing to tell anyone what it is out of spite. I hope someone else can give this some attention to resolve the problem