@kaikreuzer 2021-01-23
@spacemanspiff2007 Not sure why you think that I would be the best to look into it - there are other (and imho better) experts that already expressed their opinion [here](https://github.com/openhab/openhab-core/pull/2030#discussion_r550916552), [here](https://github.com/openhab/openhab-core/pull/2030#issuecomment-757576201) and [here](https://github.com/openhab/openhab-core/pull/2101#issuecomment-762514051). Please don't expect me to overrule their verdict in any way.

Besides this, I usually only start looking at PRs once they have some content in their description and pass the automatic checks.