@spacemanspiff2007 2021-01-24
@kaikreuzer 
I have addressed both of the first two concerns with this commit and they not longer apply.
These are not the same changes and I should have probably named this PR different than the first one.
I just want OH3 to work with basic auth again an my concern is that this PR somehow does not make it into OH3,1.


> Besides this, I usually only start looking at PRs once they have some content in their description and pass the automatic checks.

Fair enough - I added a small description what this does. 
However I have signed the DCO and all commits properly but the checks seem to fail because of the git privacy guard.
Also the automatic build failed (imho) because of another PR (``Update Mockito, JUnit and use MockitoExtensions in itests``) and not because of these changes. I do not know how I can trigger a build that only includes the changes from this PR.