@kaikreuzer 2021-01-25
> I have addressed both of the first two concerns with this commit and they not longer apply.

Is this confirmed by @wborn and @ghys? Sorry, I didn't follow in all details, but I I saw @ghys summary "tl;dr: **use API Tokens.**"
Did you consider this advice? Is there anything wrong with it? Unless there's indeed an issue, I'd also think that API tokens are the proper way to have external systems authenticate themselves against openHAB. And in the case of an NGINX as reverse-proxy, it should also be simple to add an API token to the HTTP header when passing the request to the proxy instead of turning off auth for admin endpoints altogether.