@kaikreuzer 2021-01-25
@spacemanspiff2007 And please don't get me wrong: I don't want to block you here and I am honestly looking for a solution - but I have learned to (a) listen to the experts, (b) don't mess with cryptography and (c) to really need to understand the real issue that is being tried to solve in order to think out of the box to find the best solution and not just the nearest.