@Pedals2Paddles 2021-01-26
Hey guys.  Looks like a cause and solution has been identified.  [See this thread starting here.](https://community.openhab.org/t/oh3-with-nginx-reverse-proxy-and-authentication/106528/48?u=pedals2paddles)  There is a misplaced directive in the NGINX configuration file causing this mess.  Once you move the directive down to the location/ block where it belongs, everything works beautifully. 

My configuration had the `proxy_set_header Authorization ""` directive up in the top section, not under the location/ block. This is where the OpenHabian-config tool put it twice, and where it seems to have put it for everyone else having this problem.  It is also not in the documentation on the web (which I know needs to be updated) so unlikely anyone would know or notice something is out of place.

I moved the directive down to the location/ block as suggested by ysc and mstormi. Everything now seems to operate with lightning efficiency. The CPU is no longer melting down when trying to login. It actually seems just as fast as when connected locally not through NGINX. So this certainly seems to have been the fix we’ve all been looking for.

If this takes care of it for @spacemanspiff2007 and the others having this issue, I believe this PR and associated issues can be closed with thanks.